ref: implement delegate pattern on location service
create partureScheduleViewModel to separate service logic and domain logic
